+++ CHANGELOG.md
@@ -20,6 +20,7 @@
     * New: Add `get_filter_names()` as a helper method for getting the default and third-party filter names
     * New: Add `unhook_default_filters()` as a helper method to unhook plugin's default filters hooked to `text_hover()`
     * New: Add test case for hover text that includes HTML
+    * New: Add tests for `enqueue_scripts()`
     * New: Add test for setting name
     * New: Add tests for setting defaults
     * New: Add explicit tests to ensure falsey hover text values don't alter original text